RangeTrees

Documentation for RangeTrees.

RangeTrees.RangeTree

RangeTrees._addchildren!

RangeTrees.intersect!

RangeTrees.midrange

RangeTrees.RangeTree — Type

RangeTree

An augmented, balanced, binary interval tree of ranges (or intervals) represented as UnitRange{@ref}.

The nodes field, a vector of RangeNode{@ref}s, represents the tree. The root node of a RangeTree rt is rt.nodes[rt.rootind]. The rootind field should always be midrange(UnitRange(eachindex(nodes))).

A RangeTree is usually constructed from a Vector{UnitRange}.

RangeTrees._addchildren! — Method

_addchildren!(v::Vector{RangeNode{T}}, inds::UnitRange) where T

Recursively re-write the left, right, and maxlast fields in v[inds] to form an augmented, balanced, binary interval tree.

Can be extended to store parents in the nodes.

RangeTrees.intersect! — Method

intersect!(result::AbstractVector{UnitRange}, target::UnitRange, rt::RangeTree, index)

Recursively intersect target with the intervals in the subtree of rt.nodes[index]. Non-empty intersections are push!'d onto result in sorted order.

RangeTrees.midrange — Method

midrange(rng::UnitRange{T})::T

Return the mid-value of rng rounding up when length(rng) is even.