RangeTrees

Documentation for RangeTrees.

RangeTrees.RangeNode

RangeTrees.RangeTree

Base.intersect!

RangeTrees._addchildren!

RangeTrees.midrange

RangeTrees.RangeNode — Type

RangeNode{T}

The element type of the vector in the nodes field of a RangeTree.

The value of a RangeNode{T} is the UnitRange{T} in its intvl field. The left and right fields are the indices of the children of this node (values of 0 imply no child on that side).

The maxlast field is the maximum value of last(n.intvl) for any node n in the right subtree. This is the "augmentation" in an augmented interval tree that allows for efficient intersection of an interval with all the nodes of a RangeTree.

RangeTrees.RangeTree — Type

RangeTree

An augmented, balanced, binary interval tree of intervals of integers represented as a UnitRange.

The nodes field, a vector of RangeNodes, is the tree. The root node of a RangeTree has index midrange(eachindex(nodes)).

A RangeTree is usually constructed from a Vector{UnitRange{<:Integer}}. It is designed to allow for fast intersection of another UnitRange{<:Integer} with the values of all the nodes in the tree.

```jldoctest julia> rt = RangeTree([0:0, 3:40, 10:14, 20:35, 29:98]); # example from Wikipedia page

julia> intersect(40:59, rt) 2-element Vector{UnitRange{Int64}}: 40:40 40:59 ````

Base.intersect! — Method

intersect!(result::AbstractVector{UnitRange}, target::UnitRange, rt::RangeTree, idx::Integer)

Recursively intersect target with the intervals in the subtree of rt[idx].

Non-empty intersections are pushed onto result in the same order as the intersecting nodes appear in the tree. Storing maxlast allows for the pre-order depth-first search to be truncated when a node's maxlast is less than first(target). Because the nodes are in non-decreasing order of first(intvl) the right subtree can be skipped when last(target) < first(intvl).

RangeTrees._addchildren! — Method

_addchildren!(v::Vector{RangeNode{T}}, inds::UnitRange) where T

Internal utility to recursively re-write the left, right, and maxlast fields in v[inds] so as to form an augmented, balanced, binary interval tree.

RangeTrees.midrange — Method

midrange(rng::AbstractUnitRange{T}) where {T<:Integer}

Return the median of rng, rounding up when length(rng) is even.