RangeTrees

Documentation for RangeTrees.

RangeTrees.RangeNode

RangeTrees.RangeTree

RangeTrees._addchildren!

RangeTrees.intersect!

RangeTrees.midrange

RangeTrees.RangeNode — Type

RangeNode{T}

The element type of the vector in the nodes field of a RangeTree.

The value of a RangeNode{T} is the UnitRange{T} in its intvl field. The left and right fields are the indices of the children of this node (values of 0 imply no child on that side).

The maxlast field is the maximum value of last(n.intvl) for any node n in the right subtree. This is the "augmentation" in an augmented interval tree that allows for efficient intersection of an interval with all the nodes of a RangeTree.

RangeTrees.RangeTree — Type

RangeTree

An augmented, balanced, binary interval tree of intervals of integers represented as a UnitRange.

The nodes field, a vector of RangeNodes, is the tree. The root node of a RangeTree rt is rt.nodes[rt.rootind]. The rootind field should always be midrange(UnitRange(eachindex(nodes))).

A RangeTree is usually constructed from a Vector{UnitRange{<:Integer}}. It is designed to allow for fast intersection of another UnitRange{<:Integer with the values of all the nodes in the tree.

```jldoctest julia> rt = RangeTree([0:0, 3:40, 10:14, 20:35, 29:98]); # example from Wikipedia page

julia> intersect(40:59, rt) 2-element Vector{UnitRange{Int64}}: 40:40 40:59 ````

RangeTrees._addchildren! — Method

_addchildren!(v::Vector{RangeNode{T}}, inds::UnitRange) where T

Internal utility to recursively re-write the left, right, and maxlast fields in v[inds] so as to form an augmented, balanced, binary interval tree.

RangeTrees.intersect! — Method

intersect!(result::AbstractVector{UnitRange}, target::UnitRange, rt::RangeTree, index)

Recursively intersect target with the intervals in the subtree of rt.nodes[index]. Non-empty intersections are pushed onto result in sorted order.

RangeTrees.midrange — Method

midrange(rng::UnitRange{T})::T

Return the largest median of rng.